+++ CHANGELOG.md
@@ -21,7 +21,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - HTML parser: follow and paginate links have full support of matchers, extractors and filters
 - HTTP loader: cache options are prefixed by an underscore for the sake of consistency
 - HTML parser: paginate options are prefixed by an underscore for the sake of consistency
-- CSV transformer: Jason-specific options are prefixed by an underscore for the sake of consistency
+- CSV transformer: options are split between "csv" and other file-related options 
 - E-mail transformer: options are split between "smtp" and "message"
 - All processors now received their configuration object within an object: { config }